requirements.txt
boto3==1.26.118
botocore==1.29.118
cffi==1.15.1
cryptography==40.0.2
jmespath==1.0.1
Pillow==9.5.0
pycparser==2.21
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
tk==0.1.0
urllib3==1.26.15
